LPS-54824 Move CustomAttributesDisplay classes to journal-web module, since they are registered by portletId
